#72 split test classes into packingItem-tests and packingList-tests
